requirements.txt
datasets~=2.9.0
transformers~=4.26.1
sacrebleu
evaluate~=0.4.0
rouge_score
wandb~=0.13.10
numpy~=1.24.2
tqdm~=4.64.1
torch~=1.13.1
pyyaml~=6.0
accelerate